Refactored ensemble file and added refinement hash logic

+++ source/predict.h
@@ -10,7 +10,7 @@
 apta_node* single_step(apta_node* n, tail* t, apta* a);
 double compute_score(apta_node* next_node, tail* next_tail);
 [[maybe_unused]] double predict_trace(state_merger* m, trace* tr);
-int predict_trace_type(state_merger* m, trace* tr);
+double predict_trace_type(state_merger* m, trace* tr);
 int majority_vote(std::vector<int> predictions);
 
 [[maybe_unused]] void predict(state_merger* m, inputdata& idat, std::ofstream& output);
